requirements.txt
flask==3.0.0
werkzeug==3.0.1
python-docx==1.1.0
pypdf==4.0.1
google-generativeai==0.8.0
python-dotenv==1.2.1
